[Other] Seperate option from backends (#1048)
* Seperate option from backends

* Seperate option from backends

* Seperate option from backends

* Seperate option from backends
