+++ .pre-commit-config.yaml
@@ -24,14 +24,14 @@ repos:
         files: \.(c|cc|cxx|cpp|cu|h|hpp|hxx|proto|py)$
         exclude: (?!.*third_party)^.*$
 
-# -   repo: local
-#     hooks:
-#     -   id: clang-format-with-version-check
-#         name: clang-format
-#         description: Format files with ClangFormat.
-#         entry: bash .clang_format.hook -i
-#         language: system
-#         files: \.(c|cc|cxx|cpp|cu|h|hpp|hxx|proto)$
+-   repo: local
+    hooks:
+    -   id: clang-format-with-version-check
+        name: clang-format
+        description: Format files with ClangFormat.
+        entry: bash .clang_format.hook -i
+        language: system
+        files: \.(c|cc|cxx|cpp|cu|hxx|proto)$
 
 -   repo: local
     hooks:
