+++ fastdeploy/backends/openvino/ov_backend.h
@@ -21,17 +21,10 @@
 
 #include "fastdeploy/backends/backend.h"
 #include "fastdeploy/utils/unique_ptr.h"
+#include "fastdeploy/backends/openvino/option.h"
 #include "openvino/openvino.hpp"
 
 namespace fastdeploy {
-
-struct OpenVINOBackendOption {
-  std::string device = "CPU";
-  int cpu_thread_num = -1;
-  int num_streams = 0;
-  std::map<std::string, std::vector<int64_t>> shape_infos;
-  std::set<std::string> cpu_operators{"MulticlassNms"};
-};
 
 class OpenVINOBackend : public BaseBackend {
  public:
