+++ fastdeploy/backends/tensorrt/trt_backend.h
@@ -25,6 +25,7 @@
 #include "NvOnnxParser.h"
 #include "fastdeploy/backends/backend.h"
 #include "fastdeploy/backends/tensorrt/utils.h"
+#include "fastdeploy/backends/tensorrt/option.h"
 #include "fastdeploy/utils/unique_ptr.h"
 
 class Int8EntropyCalibrator2 : public nvinfer1::IInt8EntropyCalibrator2 {
@@ -60,25 +61,6 @@ struct TrtValueInfo {
   std::vector<int> shape;
   nvinfer1::DataType dtype;   // dtype of TRT model
   FDDataType original_dtype;  // dtype of original ONNX/Paddle model
-};
-
-struct TrtBackendOption {
-  std::string model_file = "";   // Path of model file
-  std::string params_file = "";  // Path of parameters file, can be empty
-  // format of input model
-  ModelFormat model_format = ModelFormat::AUTOREC;
-
-  int gpu_id = 0;
-  bool enable_fp16 = false;
-  bool enable_int8 = false;
-  size_t max_batch_size = 32;
-  size_t max_workspace_size = 1 << 30;
-  std::map<std::string, std::vector<int32_t>> max_shape;
-  std::map<std::string, std::vector<int32_t>> min_shape;
-  std::map<std::string, std::vector<int32_t>> opt_shape;
-  std::string serialize_file = "";
-  bool enable_pinned_memory = false;
-  void* external_stream_ = nullptr;
 };
 
 std::vector<int> toVec(const nvinfer1::Dims& dim);
